Convert to workunit-focus analysis
Previous incarnation looked at the logical files generally stored on a cluster, optionally filtered by a filename pattern.  If many files made it through the filter than the cluster could be overwhelmed by metadata information requests.

This new version is based on a workunit ID.  You supply the workunit ID and the code will determine which files were used as input, then analyze those files.

Because the list of files is now pretty much guaranteed to be manageable, the code can now be turned into a query and reused as needed.

Signed-off-by: Dan S. Camper <dan.camper@lexisnexisrisk.com>